% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModStan-package.R
\docType{package}
\name{ModStan-package}
\alias{ModStan-package}
\alias{ModStan}
\title{The 'ModStan' package.}
\description{
Dose-response modelling via Rstan.
}
\details{
ModStan is developed to automate the implementation of the dose-response modelling
for analyzing phase II trials with multiple subgroups.
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.17.3.
http://mc-stan.org
}
\author{
Burak Kuersad Guenhan <burak.gunhan@med.uni-goettingen.de>
}
