% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.mod_stan}
\alias{print.mod_stan}
\title{Print mod_stan object}
\usage{
\method{print}{mod_stan}(x, digits = 2, ...)
}
\arguments{
\item{x}{A \code{mod_stan} object.}

\item{digits}{An integer indicating the number of decimal places.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The return value is invisible \code{NULL}
}
\description{
Takes an \code{mod_stan} object which is obtained by function \code{mod_stan} and print
the model and data information such as model type used in the model.
}
\details{
The resulting data.frame can be used as data argument in \code{mod_stan}.
}
