% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_stan.R
\name{mod_stan}
\alias{mod_stan}
\title{Dose-response modeling using Stan}
\usage{
mod_stan(
  dose = NULL,
  frequency = NULL,
  subgroup = NULL,
  resp = NULL,
  reference_freq = NULL,
  data = NULL,
  N_pred = 30,
  model = "Shrinkage",
  beta_prior = 0.5,
  beta_prior_dist = "half-normal",
  chains = 3,
  iter = 2000,
  warmup = 1000,
  adapt_delta = 0.95
)
}
\arguments{
\item{dose}{A numerical vector specfiying the dose values}

\item{frequency}{A numerical vector specfiying the frequency of adminsitration in hours.
This is only needed when subgroups refer to dose regimens.}

\item{subgroup}{A numerical vector specfiying the subgroup indicator}

\item{resp}{A numerical vector specfiying the response values}

\item{reference_freq}{A numerical value specifying the frequency of administration in hours
for the reference dose regimen.}

\item{data}{Optional data frame containing the variables given to the arguments above}

\item{model}{A string specifying the model used. Available options are `Pooling`,
`Stratified`, and `Shrinkage`. Default is `Shrinkage`.}

\item{beta_prior}{A numerical value specifying the standard deviation of the prior density
for heterogenety stdev. Default is 0.5.}

\item{beta_prior_dist}{A string specifying the prior density for the heterogeneity standard deviation,
option is `half-normal` for half-normal prior, `uniform` for uniform prior, `half-cauchy` for
half-cauchy prior.}

\item{chains}{A positive integer specifying the number of Markov chains.
The default is 4.}

\item{iter}{A positive integer specifying the number of iterations for each chain
(including warmup). The default is 2000.}

\item{warmup}{A positive integer specifying the number of warmup (aka burnin)
iterations per chain. The default is 1000.}

\item{adapt_delta}{A numerical value specfying the target average proposal acceptance
probability for adaptation. See Stan manual for details. Default is 0.95. In general
you should not need to change adapt_delta unless you see a warning message about
divergent transitions, in which case you can increase adapt_delta from the
default to a value closer to 1 (e.g. from 0.95 to 0.99, or from 0.99 to 0.999, etc).}
}
\value{
an object of class `stanfit` returned by `rstan::sampling`
}
\description{
`mod_stan` fits a dose-response model with subgroups using Stan. Currently, only "emax"
model is available.
}
\examples{
\donttest{
data('dat.Dupilumab', package = "ModStan")
## Fitting an Emax model using shrinkage estimation for ED50 parameters
partial.shrinkage.Dupilumab.stan = mod_stan(dose = dose,
                                           resp = resp,
                                           subgroup = subgroup,
                                           frequency = frequency,
                                           data = dat.dupilumab,
                                           model = "Shrinkage",
                                           reference_freq = 24 * 7 * 8,
                                           beta_prior_dist = "half-normal",
                                           beta_prior = 0.5,
                                           chains = 4,
                                           iter = 2000,
                                           warmup = 1000)
## Obtaining a small summary
print(shrinkage.Dupilumab.stan$fit, pars = c("theta_0", "theta_1", "theta_2", "beta_theta_2", "sigma"))
## Extract the rstan fit for post-processing, eg convergence diagnostics
shrinkage.dupilumab.stanfit = partial.shrinkage.Dupilumab.stan$fit_sum

}

}
